// pijamas/src/types/database.ts
export type ProductModel = 'conjunto' | 'camisola' | 'short_doll' | 'baby_doll' | 'roupao'
export type ProductFabric = 'algodao' | 'seda' | 'cetim' | 'fleece' | 'malha'
export type ProductSize = 'PP' | 'P' | 'M' | 'G' | 'GG' | '2XG' | '3XG' | '2' | '4' | '6' | '8' | '10' | '12' | '14'
export type ProductStatus = 'disponivel' | 'reservado' | 'esgotado' | 'sob_encomenda'